LayoutTests/fast/loader/url-selected-user-gesture.html

<!DOCTYPE html>
<script src="../js/resources/js-test-pre.js"></script>
<a href='about:blank' target='_blank'></a>
<script>
    description('Check that a javascript generated mouse click is not enough to open a new tab.');
    if (window.testRunner) {
        testRunner.setCanOpenWindows();
        testRunner.setPopupBlockingEnabled(true);
    }

    document.querySelector('a').dispatchEvent(new MouseEvent('click', {ctrlKey: true, metaKey: true}));

    if (window.testRunner)
        shouldBe("testRunner.windowCount()", "1");
</script>
<script src="../js/resources/js-test-post.js"></script>
